services/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
pydantic>=2.8,<3.0
pydantic-settings>=2.4,<3.0
PyJWT>=2.9,<3.0
argon2-cffi>=23.1,<24.0
email-validator>=2.2,<3.0
google-genai>=0.3,<1.0
pypdf>=5.0,<6.0
httpx>=0.27,<1.0
pytest>=8.3,<9.0
